--- a/src/Roave/InfectionStaticAnalysis/Psalm/RunStaticAnalysisAgainstMutant.php
+++ b/src/Roave/InfectionStaticAnalysis/Psalm/RunStaticAnalysisAgainstMutant.php
@@ -36,6 +36,9 @@ class RunStaticAnalysisAgainstMutant
                 ->getOriginalFilePath()
         );
 
+        $codebase->addFilesToAnalyze([$path => $path]);
+        $codebase->scanFiles();
+
         if (! $this->alreadyVisitedStubs) {
             $codebase->config->visitPreloadedStubFiles($codebase);
             $codebase->config->visitStubFiles($codebase);