--- a/test/asset/MutableProjectStub/psalm.xml
+++ b/test/asset/MutableProjectStub/psalm.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<psalm
+    xmlns="https://getpsalm.org/schema/config"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="https://getpsalm.org/schema/config ../../../vendor/vimeo/psalm/config.xsd"
+    name="Static analysis stubs verifying that stub mutations are not picked up"
+>
+    <projectFiles>
+        <directory name="."/>
+    </projectFiles>
+</psalm>