--- a/test/asset/PreloadClassStub/Stub.phpstub
+++ b/test/asset/PreloadClassStub/Stub.phpstub
@@ -1,0 +1,5 @@
+<?php
+
+namespace Roave\InfectionStaticAnalysisAsset\PreloadClassStub;
+
+interface Stub {}