Fixes #4

Quoting original issue:

 > As discussed with @weirdan and @muglug on https://github.com/vimeo/psalm/issues/4117#issuecomment-686759090, there is a more efficient way to run static analysis repeatedly on a codebase:
 >
 > https://github.com/vimeo/psalm/blob/8c7423505a4282d1cdd38f3ff9552b911cbd5fa9/tests/FileUpdates/ErrorFixTest.php#L53
 >
 > In practice, `Codebase#reloadFiles()` should cause a file reload of only what's relevant to us, allowing us to avoid a complete cache bust like we do in https://github.com/Roave/infection-static-analysis-plugin/blob/5bb581368e244cbf7035d65faad2112e5547868c/bin/roave-infection-static-analysis-plugin#L40-L52

## Blockers

 * [ ] https://github.com/vimeo/psalm/issues/5764