@Ocramius 2021-05-13
Rough comparison:

### Before

```
Time: 8s. Memory: 0.04GB

real	0m9,182s
user	0m8,148s
sys	0m1,180s
```

### After

```
Time: 7s. Memory: 18.00MB

real	0m8,205s
user	0m7,124s
sys	0m1,289s
```

Note how memory usage decreased by almost twice.

Trying it on a larger project before calling it a victory.