@Ocramius 2021-05-13
On a much larger codebase:

### Before

```
time php -dmemory_limit=-1 vendor/bin/roave-infection-static-analysis-plugin 

    ____      ____          __  _
   /  _/___  / __/__  _____/ /_(_)___  ____
   / // __ \/ /_/ _ \/ ___/ __/ / __ \/ __ \
 _/ // / / / __/  __/ /__/ /_/ / /_/ / / / /
/___/_/ /_/_/  \___/\___/\__/_/\____/_/ /_/

Infection - PHP Mutation Testing Framework version 0.23.0@ac87ae49e9f192f610276565e54469231dca7837


Running initial test suite...

PHPUnit version: 9.5.2

  350 [============================] 5 secs

Generate mutants...

Processing source code files: 184/460^C
real	0m9.881s
user	0m9.376s
sys	0m0.536s
docker@0d20b3344a1f:/var/www/kanbanbox/locapp.kanbanbox.com$ 
docker@0d20b3344a1f:/var/www/kanbanbox/locapp.kanbanbox.com$ time php -dmemory_limit=-1 vendor/bin/roave-infection-static-analysis-plugin 

    ____      ____          __  _
   /  _/___  / __/__  _____/ /_(_)___  ____
   / // __ \/ /_/ _ \/ ___/ __/ / __ \/ __ \
 _/ // / / / __/  __/ /__/ /_/ / /_/ / / / /
/___/_/ /_/_/  \___/\___/\__/_/\____/_/ /_/

Infection - PHP Mutation Testing Framework version 0.23.0@ac87ae49e9f192f610276565e54469231dca7837


Running initial test suite...

PHPUnit version: 9.5.2

  349 [============================] 5 secs

Generate mutants...

Processing source code files: 195/195
.: killed, M: escaped, U: uncovered, E: fatal error, T: timed out, S: skipped

UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   ( 50 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   (100 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   (150 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   (200 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU.......   (250 / 963)
........MMMMM..MMMMM....MMMMM...MMMMM...MMMMMMMMM.   (300 / 963)
............TT............MMMMM...MMMMM.M.MMMMMMMM   (350 / 963)
M.................MMMMM................M........M.   (400 / 963)
..................................................   (450 / 963)
..................................................   (500 / 963)
......................MM.............M.M.....M....   (550 / 963)
M............MM............M......................   (600 / 963)
........M...M.....................................   (650 / 963)
............................M.M...................   (700 / 963)
.......................M.M........................   (750 / 963)
..................................................   (800 / 963)
..................................................   (850 / 963)
..................................................   (900 / 963)
..................................................   (950 / 963)
.............                                        (963 / 963)

963 mutations were generated:
     647 mutants were killed
     243 mutants were not covered by tests
      71 covered mutants were not detected
       0 errors were encountered
       2 time outs were encountered
       0 mutants required more time than configured

Metrics:
         Mutation Score Indicator (MSI): 67%
         Mutation Code Coverage: 74%
         Covered Code MSI: 90%

<SNIP>

Time: 5m 33s. Memory: 0.12GB

real	5m36.237s
user	3m40.995s
sys	0m59.706s
```

### After

```
time php -dmemory_limit=-1 vendor/bin/roave-infection-static-analysis-plugin 

    ____      ____          __  _
   /  _/___  / __/__  _____/ /_(_)___  ____
   / // __ \/ /_/ _ \/ ___/ __/ / __ \/ __ \
 _/ // / / / __/  __/ /__/ /_/ / /_/ / / / /
/___/_/ /_/_/  \___/\___/\__/_/\____/_/ /_/

Infection - PHP Mutation Testing Framework version 0.23.0@ac87ae49e9f192f610276565e54469231dca7837


Running initial test suite...

PHPUnit version: 9.5.2

  349 [============================] 6 secs

Generate mutants...

Processing source code files: 195/195
.: killed, M: escaped, U: uncovered, E: fatal error, T: timed out, S: skipped

UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   ( 50 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   (100 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   (150 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU   (200 / 963)
UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUU.......   (250 / 963)
........M.........................................   (300 / 963)
............TT....................................   (350 / 963)
..................................................   (400 / 963)
..................................................   (450 / 963)
..................................................   (500 / 963)
......................MM.............M.M..........   (550 / 963)
M..........................M......................   (600 / 963)
..................................................   (650 / 963)
..................................................   (700 / 963)
..................................................   (750 / 963)
..................................................   (800 / 963)
..................................................   (850 / 963)
..................................................   (900 / 963)
..................................................   (950 / 963)
.............                                        (963 / 963)

963 mutations were generated:
     711 mutants were killed
     243 mutants were not covered by tests
       7 covered mutants were not detected
       0 errors were encountered
       2 time outs were encountered
       0 mutants required more time than configured

Metrics:
         Mutation Score Indicator (MSI): 74%
         Mutation Code Coverage: 74%
         Covered Code MSI: 99%

Time: 3m 39s. Memory: 0.10GB

real	3m42.681s
user	2m2.229s
sys	0m42.430s
```

The times are much improved, **BUT** there is a major difference that invalidates the results:

```diff
-     647 mutants were killed
+     711 mutants were killed
     243 mutants were not covered by tests
-      71 covered mutants were not detected
+       7 covered mutants were not detected
```

This means that the static analyzer has changed behavior due to this patch, potentially failing on cross-file dependencies. I think we need a more careful test that is based on cross-file dependencies (stubs loaded by the analyzer).

In theory, we can obtain this massive performance improvement while also retaining the stability, but it needs work and more careful analysis of the explanation in #4.