@Ocramius 2021-05-13
After some investigation work, what went wrong came out clearly: psalm detects `DuplicateClass` for each mutant after the first one, whenever more than one mutant escapes per class.

The test case to be written is therefore:

 * a class or function in a file
 * multiple escaping mutant supposed to be detected for it

This test can probably be performed by attempting static analysis on multiple escaped mutants with the same source contents (duplicate symbols) in https://github.com/Roave/infection-static-analysis-plugin/blob/c1e5428a85ff24fa4ca75ac43f29c608d02379fe/test/unit/Roave/InfectionStaticAnalysisTest/Psalm/RunStaticAnalysisAgainstMutantTest.php